hwcomposer: Invalidate whole buffer cache on HWC_GEOMETRY_CHANGED
HWC contract guarantees that a layer buffer can be reused,
but only if it didn't change on two consecutive prepare calls
*and* the HWC_GEOMETRY_CHANGED flag is not set. With the way
buffer cache worked so far, it was possible for an old buffer
to get reused after HWC_GEOMETRY_CHANGED was set if it wasn't
present in the layer list that had that flag set.

This fixes flickering in subsurface mode.