hwcomposer: Delay closing window to next hwc_set
When the compositor asks us to close the window we ask the activity
service to quit, then immediatly destroy the window and its surfaces.
However, the next time hwc_set is called the application might still
be running, which causes a crash when trying to commit the already
destroyed surface.
Since hwc_set already handles closing windows when apps are closed
from the android world, reuse that logic to destroy the window.