hwcomposer: Use EGL to update shmem buffers from GPU
Surfaceflinger layers are not created with SW_READ flags so we should
not be mapping them in main memory.
It still works for gralloc.default buffers because they're already
implemented with ashmem.

This code path is used when the we can use GPU buffers but the
compositor does not implement the dmabuf protocol.