hwcomposer: Remove broken timedwait on size data
The timedwait never worked for multiple reasons, two of which are:
* Bad arguments prior to f90f8aa
* When calibrating we are not running a separate wayland thread yet
  so everything is synchronized. If we had not received the data
  we wanted, no amount of sleeping this thread could fix it.

Unfortunately in commit f90f8aa I accidentally put the timedwait
in a while-loop meaning we would never advance.

wlroots based compositors seem to not be sending the size data we
need. We had a fallback mechanism in place, so continue using that.

Closes: #53