[AudioHAL] Pass frames instead of bytes to ALSA `set_buffer_size()`
According to the ALSA documentation [0], this method takes frames in as
the third parameter, however our `out_get_buffer_size()` function
returns bytes.  We do the conversion here to ensure that both sides
agree on the relevant buffer sizes.

Note that in our experiments, we were able to achieve significantly
lower latency with this patch and by reducing `PLAYBACK_PERIOD_SIZE` to
`512`, however we recognize this may not be desirable on all hardware,
and so have omitted that change here.

[0] https://www.alsa-project.org/alsa-doc/alsa-lib/group___p_c_m___h_w___params.html#ga828cf9e11fe84137c2c8d4938a6fca7f