hwcomposer: Use hardware cursor even in full-ui or single-app mode
Use the HWC_IS_CURSOR_LAYER flag to detect the cursor layer and request a
compositionType HWC_OVERLAY so that we can handle the cursor layer
independently and set it as the wayland cursor surface.

This is only now possible because we added support for the cursor hotspot,
otherwise we would be degrading the experience.