hwcomposer: Work around a deadlock
When the hwc_set thread coincidentally dispatches a xdg_surface_close event while it is doing its roundtrip in window::create, it will deadlock since it tries to lock windowsMutex which it already holds. Work around this by using a recursive_mutex.
Note: It is still possible that the wayland thread will lock up in xdg_surface_close in this case but since the hwc_set thread still dispatches events and thus will not lock up, it will release the windowsMutex eventually