hwcomposer: Select cursor mode before creating the first window
Because we do roundtrips for the configure events, we might end up handling
pointer events in window::create.

Make sure the cursor handler is set before creating the first window.
This is safe because create_display also runs a roundtrip for the global
registry events, where we bind the protocols needed by the wl_cursor handler.