hwcomposer: Calibrate window dimensions from the last configure event
Hyprland is sending an initial configure event with a valid size, but it is
not the proper size accounting for the window margins; it only sends the real
desired size after a buffer is attached.

Rework calibration to honour the last configure event after all the roundtrips.

Hopefully all compositors that were previously sending valid dimensions on the
first configure event are nice and keep sending valid dimensions after
attaching a 1x1 buffer.
If that turns out not to be the case, we could also try configuring the
viewporter to the dimensions requested with the first configure event.