hwcomposer: Lock pointer on every layer surface
The main window->surface is generally either a solid background surface, or the
"empty" background surface used for multi-windows mode.

When the pointer is on of the overlapping layer subsurfaces, the pointer lock
wouldn't be active. We want to lock on an actual app layer surface instead.

This fixes a regression introduced with https://github.com/waydroid/android_hardware_waydroid/pull/63,
and also makes pointer locking work in multi-windows mode.

Closes: https://github.com/waydroid/waydroid/issues/2006